requirements.txt
fastapi
uvicorn
aiogram>=3.17.0
openai
mcp
python-dotenv
